# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LessonsRootIn": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "LessonsRootIn",
        "type": "object"
      },
...
      "ModelIn": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "model"
        ],
        "title": "ModelIn",
        "type": "object"
      },
...
      "SecretIn": {
        "properties": {
          "value": {
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "SecretIn",
        "type": "object"
      },
      "SecretSaved": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "set": {
            "default": true,
            "title": "Set",
            "type": "boolean"
          },
          "stored_in": {
            "enum": [
              "store",
              "env"
            ],
            "title": "Stored In",
            "type": "string"
          }
        },
        "required": [
          "name",
          "stored_in"
        ],
        "title": "SecretSaved",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token mostrato da 'rt api' al primo avvio",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nAPI locale di RT (utente singolo). Autenticazione: premi **Authorize** e incolla il token\nmostrato da `rt api` al primo avvio (header `Authorization: Bearer <token>`). Per un token\nnuovo: `rt api --reset-token`.\n\nLe operazioni lunghe (pipeline, fasi, trascrizione, immagini, recall) diventano job eseguiti\nda `rt worker`: l'endpoint risponde 202 con l'id del job, il progresso arriva da\n`GET /jobs/{id}/events` (Server-Sent Events).\n",
    "title": "RT API",
    "version": "rt"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/secrets/{name}": {
      "put": {
        "operationId": "put_secret_api_v1_secrets__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Salva un segreto dichiarato (archivio cifrato se inizializzato); il valore non viene mai restituito",
        "tags": [
          "impostazioni"
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tutte le impostazioni (nessun valore segreto)",
        "tags": [
          "impostazioni"
        ]
      }
    },
    "/api/v1/settings/connections": {
      "post": {
        "operationId": "post_connection_api_v1_settings_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Nuova connessione LLM (provider, base URL, una o più chiavi)",
        "tags": [
          "impostazioni"
        ]
      }
    },
    "/api/v1/settings/connections/{name}/models": {
      "post": {
        "operationId": "post_connection_model_api_v1_settings_connections__name__models_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aggiunge un modello a una connessione",
        "tags": [
          "impostazioni"
        ]
      }
    },
    "/api/v1/settings/lessons-root": {
      "put": {
        "operationId": "put_lessons_root_api_v1_settings_lessons_root_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonsRootIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cartella delle lezioni",
        "tags": [
          "impostazioni"
        ]
      }
    },
    "/api/v1/settings/models/test": {
      "post": {
        "operationId": "post_model_test_api_v1_settings_models_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelTestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelTestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Autenticazione mancante o non valida"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CSRF non valido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Risorsa non trovata"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflitto (es. job in corso sulla lezione)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Richiesta non valida"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Prova connessione e modello con una chiamata minima (anche prima di salvarli)",
        "tags": [
          "impostazioni"
        ]
      }
    },
...
  }
}